# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CodeMind API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/lessons/questions/[questionId]": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "summary": "API endpoint for /api/lessons/questions/[questionId]",
        "description": "Tự động sinh tài liệu cho GET /api/lessons/questions/[questionId]. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/questions/[questionId]": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "API endpoint for /api/questions/[questionId]",
        "description": "Tự động sinh tài liệu cho GET /api/questions/[questionId]. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/roadmap/node-detail": {
      "post": {
        "tags": [
          "Roadmap AI"
        ],
        "summary": "Tạo chi tiết cho một Node (bài học) bằng AI",
        "description": "Gọi API Groq để phân tích và sinh ra nội dung giải thích, tài liệu học tập cho một chủ đề cụ thể.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "topic"
                ],
                "properties": {
                  "topic": {
                    "type": "string",
                    "example": "React Hooks"
                  },
                  "context": {
                    "type": "string",
                    "example": "Lập trình Frontend cơ bản"
                  },
                  "user_level": {
                    "type": "string",
                    "example": "beginner"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thông tin chi tiết của node"
          },
          "400": {
            "description": "Thiếu topic"
          },
          "500": {
            "description": "Lỗi máy chủ hoặc lỗi parse JSON"
          }
        }
      }
    },
...
  },
...
}